change test files to tsx
